Added try-with-resources on some ResultSets.
Goal was to try stopping locking errors, and so this code just started
putting ResultSet and Statement use instead try-with-resource blocks to make
sure that they automatically closed. At least informally, it
seemed to work. This was not a comprehensive change; I just changed as
many as I needed to to make error go away.